@@ -1,20 +1,14 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import App from './App';
-import { createTheme, ThemeProvider } from '@mui/material/styles';
-import { ruRU } from '@mui/material/locale';
-import { ruRU as ruRUGrid } from '@mui/x-data-grid-pro';
-const theme = createTheme(
-  {},
-  ruRU,
-  ruRUGrid
-);
-
-ReactDOM.render(
+import React from "react";
+import ReactDOM from "react-dom/client";
+import App from "./App";
+import { createTheme, ThemeProvider } from "@mui/material/styles";
+import { ruRU } from "@mui/material/locale";
+import { ruRU as ruRUGrid } from "@mui/x-data-grid-pro";
+const theme = createTheme({}, ruRU, ruRUGrid);
+ReactDOM.createRoot(document.getElementById("root")).render(
   <React.StrictMode>
     <ThemeProvider theme={theme}>
-        <App />
+      <App />
     </ThemeProvider>
-  </React.StrictMode>,
-  document.getElementById('root')
+  </React.StrictMode>
 );
